requirements.txt
flask
openai
gunicorn
flask-limiter
psycopg2-binary
requests
beautifulsoup4
